Add hacky --forceExit command while we figure out what open handles are keeping jest from exiting
Before this change, the test processes hang because there are some open handles keeping node from exiting. I tried to figure out what they were with `jest --detectOpenHandles --runInBand`, but it never exited or printed any handles for me. I think we'll have to spelunk a different way, but in the meantime, let's get the tests working so we can get some signal on `main` to know if stuff is broken!