Don't run a server side fetch for different anchors at the same URL
This is a purely client side navigation where the client already has all the information necessary to render the page. We assume the contents don't change. Someday, we're gonna need a way for the application to manage the cache, we don't have that yet, but that'd be the way to tell fastify-renderer to force reload the data from the server.