Add a useNavigationDetails hook for informing userland code that a navigation is underway
This is necessary because of the way we use Suspense and useTransition for navigation. The state change for the currently rendered path only exists on the new "branch" of the render tree, and the old tree still sees the old location and old state so that it remains interactive for the user. The old state might want to show a pending indicator or what have you when there is a navigation underway, so we expose a nice hook to grab that information.