Revert "upgrade to react alpha"
This original change makes a lot of sense, but it means we can't upgrade Gadget to use the new features / fixes we're adding to fastify-renderer because it is not yet compatible with some of the changes between this experimental version and the React 18 alpha. That will end in a couple weeks at which point I will revert this commit and we'll go back to being as close to 18 as possible. Woop woop!

This reverts commit dc277b1e636595fa108a0a9cf67fb050ef167b8b.